Bugs bugs bugs...

+++ lessons/ch9/MoreShapes/Point.java
@@ -6,8 +6,8 @@
 
 public class Point
 {
-    private float x;
-    private float y;
+    private double x;
+    private double y;
 
     // Default no arguement constructor.
     public Point()
@@ -16,30 +16,30 @@ public class Point
     }
     
     // Specific Constructor (x,y)
-    public Point(float x, float y)
+    public Point(double x, double y)
     {
 	this.x = x;
 	this.y = y;
     }
 
     // Get Functions
-    public float getX()
+    public double getX()
     {
 	return x;
     }
 
-    public float getY()
+    public double getY()
     {
 	return y;
     }
 
     // Mutators
-    public void setX(float x)
+    public void setX(double x)
     {
 	this.x = x;
     }
 
-    public void setY(float y)
+    public void setY(double y)
     {
 	this.y = y;
     }
